more refactoring
