+++ tests/test_unit/test_report_a_suspected_breach/test_forms.py
@@ -332,7 +332,10 @@ class TestUploadDocumentsForm:
         assert "document" in form.errors
         assert form.errors.as_data()["document"][0].code == "too_large"
 
-    @patch("report_a_suspected_breach.forms.forms_e.get_all_session_files", return_value=MockAllSessionFiles(length=10))
+    @patch(
+        "report_a_suspected_breach.forms.forms_documents_and_details.get_all_session_files",
+        return_value=MockAllSessionFiles(length=10),
+    )
     def test_too_many_uploaded(self, mocked_get_all_session_files, request_object):
         good_file = SimpleUploadedFile("good.pdf", b"%PDF-test_pdf")
